requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt  # For token-based authentication

